Fix compilation for SDCC 4.0.0

+++ Firmware/radio/radio.h
@@ -52,7 +52,7 @@
 
 // the biggest packet length we will allow. To allow for golay
 // encoding this needs to be a multiple of 6
-#define MAX_PACKET_LENGTH 252
+#define MAX_PACKET_LENGTH ((uint8_t)252)
 
 
 #include "board.h"
